# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Medynium API",
    "description": "Governed Patient 360 and clinical agent on Snowflake. Decision support on synthetic data only. Errors use {error, message}. Sessions use HttpOnly cookies; state-changing requests need the X-Medynium-Client: web header.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/patients": {
      "get": {
        "tags": [
          "patients"
        ],
        "summary": "List Patients",
        "description": "The caller's entitled patients. `q` searches name or id; `changed` keeps only flagged patients.",
        "operationId": "list_patients_patients_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "name": "changed",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Changed"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_PatientListItem_"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/patients/{patient_id}": {
      "get": {
        "tags": [
          "patients"
        ],
        "summary": "Get Patient",
        "description": "Patient 360 overview. 404 if the patient is missing or the caller is not entitled.",
        "operationId": "get_patient_patients__patient_id__get",
        "parameters": [
          {
            "name": "patient_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Patient Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Overview"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/patients/{patient_id}/medications": {
      "get": {
        "tags": [
          "patients"
        ],
        "summary": "Medications",
        "operationId": "medications_patients__patient_id__medications_get",
        "parameters": [
          {
            "name": "patient_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Patient Id"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "active",
                "all"
              ],
              "type": "string",
              "default": "active",
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Medication"
                  },
                  "title": "Response Medications Patients  Patient Id  Medications Get"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/patients/{patient_id}/labs": {
      "get": {
        "tags": [
          "patients"
        ],
        "summary": "Labs",
        "description": "Latest and previous value per test, with reference range and flag.",
        "operationId": "labs_patients__patient_id__labs_get",
        "parameters": [
          {
            "name": "patient_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Patient Id"
            }
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LabLatest"
                  },
                  "title": "Response Labs Patients  Patient Id  Labs Get"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/patients/{patient_id}/labs/{code}/trend": {
      "get": {
        "tags": [
          "patients"
        ],
        "summary": "Lab Trend",
        "description": "`code` is a LOINC code or a short alias such as eGFR or HbA1c.",
        "operationId": "lab_trend_patients__patient_id__labs__code__trend_get",
        "parameters": [
          {
            "name": "patient_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Patient Id"
            }
          },
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Code"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LabTrend"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/patients/{patient_id}/timeline": {
      "get": {
        "tags": [
          "patients"
        ],
        "summary": "Timeline",
        "description": "Merged events, newest first. Filter by date range and comma-separated event types.",
        "operationId": "timeline_patients__patient_id__timeline_get",
        "parameters": [
          {
            "name": "patient_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Patient Id"
            }
          },
          {
            "name": "from",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "From"
            }
          },
          {
            "name": "to",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "To"
            }
          },
          {
            "name": "types",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Types"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Timeline"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/patients/{patient_id}/claims": {
      "get": {
        "tags": [
          "patients"
        ],
        "summary": "Claims",
        "description": "Utilisation and claims, each linked to its encounter.",
        "operationId": "claims_patients__patient_id__claims_get",
        "parameters": [
          {
            "name": "patient_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Patient Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Claims"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/patients/{patient_id}/notes": {
      "get": {
        "tags": [
          "patients"
        ],
        "summary": "Notes",
        "operationId": "notes_patients__patient_id__notes_get",
        "parameters": [
          {
            "name": "patient_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Patient Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/NoteSummary"
                  },
                  "title": "Response Notes Patients  Patient Id  Notes Get"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/patients/{patient_id}/notes/{note_id}": {
      "get": {
        "tags": [
          "patients"
        ],
        "summary": "Note",
        "description": "A note is returned as plain text and is treated as data, never as instructions.",
        "operationId": "note_patients__patient_id__notes__note_id__get",
        "parameters": [
          {
            "name": "patient_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Patient Id"
            }
          },
          {
            "name": "note_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Note Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NoteDetail"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Claims": {
        "properties": {
          "utilization": {
            "$ref": "#/components/schemas/Utilization"
          },
          "claims": {
            "items": {
              "$ref": "#/components/schemas/Claim"
            },
            "type": "array",
            "title": "Claims"
          }
        },
        "type": "object",
        "required": [
          "utilization",
          "claims"
        ],
        "title": "Claims"
      },
...
      "ErrorBody": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "details": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details"
          }
        },
        "type": "object",
        "required": [
          "error",
          "message"
        ],
        "title": "ErrorBody"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LabLatest": {
        "properties": {
          "lab_id": {
            "type": "string",
            "title": "Lab Id"
          },
          "test": {
            "type": "string",
            "title": "Test"
          },
          "code": {
            "type": "string",
            "title": "Code"
          },
          "value": {
            "type": "number",
            "title": "Value"
          },
          "unit": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unit"
          },
          "date": {
            "type": "string",
            "format": "date",
            "title": "Date"
          },
          "previous": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PreviousValue"
              },
              {
                "type": "null"
              }
            ]
          },
          "ref": {
            "$ref": "#/components/schemas/Reference"
          },
          "flag": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "LOW",
                  "HIGH",
                  "NORMAL"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Flag"
          },
          "source": {
            "type": "string",
            "title": "Source",
            "default": "CLINICAL.LAB_RESULT"
          }
        },
        "type": "object",
        "required": [
          "lab_id",
          "test",
          "code",
          "value",
          "unit",
          "date",
          "previous",
          "ref",
          "flag"
        ],
        "title": "LabLatest"
      },
      "LabTrend": {
        "properties": {
          "test": {
            "type": "string",
            "title": "Test"
          },
          "code": {
            "type": "string",
            "title": "Code"
          },
          "unit": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unit"
          },
          "ref": {
            "$ref": "#/components/schemas/Reference"
          },
          "points": {
            "items": {
              "$ref": "#/components/schemas/TrendPoint"
            },
            "type": "array",
            "title": "Points"
          }
        },
        "type": "object",
        "required": [
          "test",
          "code",
          "unit",
          "ref",
          "points"
        ],
        "title": "LabTrend"
      },
...
      "Medication": {
        "properties": {
          "medication_id": {
            "type": "string",
            "title": "Medication Id"
          },
          "drug": {
            "type": "string",
            "title": "Drug"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "dose": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dose"
          },
          "strength": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Strength"
          },
          "started": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started"
          },
          "stopped": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stopped"
          },
          "last_change_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Change Date"
          },
          "change": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Change"
          },
          "in_knowledge_base": {
            "type": "boolean",
            "title": "In Knowledge Base"
          },
          "also_sold_as": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Also Sold As"
          },
          "source": {
            "type": "string",
            "title": "Source",
            "default": "CLINICAL.MEDICATION"
          }
        },
        "type": "object",
        "required": [
          "medication_id",
          "drug",
          "description",
          "dose",
          "strength",
          "started",
          "stopped",
          "last_change_date",
          "change",
          "in_knowledge_base",
          "also_sold_as"
        ],
        "title": "Medication"
      },
...
      "NoteDetail": {
        "properties": {
          "note_id": {
            "type": "string",
            "title": "Note Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Type"
          },
          "date": {
            "type": "string",
            "format": "date",
            "title": "Date"
          },
          "encounter_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Encounter Id"
          },
          "author": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Author"
          },
          "body": {
            "type": "string",
            "title": "Body"
          }
        },
        "type": "object",
        "required": [
          "note_id",
          "title",
          "type",
          "date",
          "encounter_id",
          "author",
          "body"
        ],
        "title": "NoteDetail"
      },
      "NoteSummary": {
        "properties": {
          "note_id": {
            "type": "string",
            "title": "Note Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Type"
          },
          "date": {
            "type": "string",
            "format": "date",
            "title": "Date"
          },
          "encounter_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Encounter Id"
          }
        },
        "type": "object",
        "required": [
          "note_id",
          "title",
          "type",
          "date",
          "encounter_id"
        ],
        "title": "NoteSummary"
      },
      "Overview": {
        "properties": {
          "patient_id": {
            "type": "string",
            "title": "Patient Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "age": {
            "type": "integer",
            "title": "Age"
          },
          "sex": {
            "type": "string",
            "title": "Sex"
          },
          "city": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "City"
          },
          "as_of": {
            "type": "string",
            "format": "date",
            "title": "As Of"
          },
          "diagnoses": {
            "items": {
              "$ref": "#/components/schemas/Diagnosis"
            },
            "type": "array",
            "title": "Diagnoses"
          },
          "medications": {
            "items": {
              "$ref": "#/components/schemas/Medication"
            },
            "type": "array",
            "title": "Medications"
          },
          "latest_labs": {
            "items": {
              "$ref": "#/components/schemas/LabLatest"
            },
            "type": "array",
            "title": "Latest Labs"
          },
          "recent_events": {
            "items": {
              "$ref": "#/components/schemas/TimelineEvent"
            },
            "type": "array",
            "title": "Recent Events"
          },
          "utilization": {
            "$ref": "#/components/schemas/Utilization"
          },
          "agent_scope_label": {
            "type": "string",
            "title": "Agent Scope Label",
            "default": "Agent scope: this patient"
          }
        },
        "type": "object",
        "required": [
          "patient_id",
          "name",
          "age",
          "sex",
          "city",
          "as_of",
          "diagnoses",
          "medications",
          "latest_labs",
          "recent_events",
          "utilization"
        ],
        "title": "Overview"
      },
...
      "Page_PatientListItem_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/PatientListItem"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "Page[PatientListItem]"
      },
...
      "Timeline": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/TimelineEvent"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total"
        ],
        "title": "Timeline"
      },
...
    }
  }
}